[vbox] Improve power cycle before export
Do a power cycle to ensure everything is good and wait for 4 minutes to
give the internet detector time to detect the network change. A similar
code was before in `ensure_vm_running`, but the waiting time only
happened if the VM was in state `saved` when starting the shutdown. In
addition, it unnecessarily delayed other functions that call this
function. Improve also the output making it consistent with the rest of
the code.