[vbox] Rename snapshots to prevent wrong restore
The `vbox-build-flare-vm.py` script could restore the wrong snapshot if
a snapshot with the same name already existed. This would break the
intended logic, as the wrong snapshot would be used for further
processing.

This commit introduces a new function `rename_old_snapshot` which
renames existing snapshots matching a given name by appending ` OLD` to
them. The `rename_old_snapshot` function is called before taking the
base snapshot and before taking each of the final export snapshots. This
prevents ambiguity and ensures the correct snapshot is used throughout
the script's execution.